requirements.txt
numpy
pandas
scikit-learn
requests
beautifulsoup4
newspaper3k
sentence-transformers
torch
tqdm
lxml